fix ByteCountingCRC32: possible invalid counting with update(bytes, offset, length)
as soon as update(bytes, offset, length) is called with more length than
there is in the array, or the offset restricts the length even further,
counting failed with invalid values.

So some sanity checks are added to help maintain validity of the counter.

Although these corner cases should not occur in real-life, it is good
to have them addressed and ensure that no malicious code may crash the
byte counter and render it invalid.