refact: use weak references to store builders
usingw eak references and creating new builders as needed
will help to keep the memory footprint low.
So all the builders will be released as soon as the zipping
has ended. This is good in case this library is used inside a
server process and zipping takes place only occasionly.