refactor IO utilities and provide ProxyInputStream
Some more `InputStream` utilities help to deal with
counting and compressing `InputStream`s on the fly.

All these utility classes are available as part of the
[Apache commons IOUtils](https://commons.apache.org/proper/commons-io/) package.
Nevertheless we would like to keep external dependencies
to a minimum and therefore implement our own version
for simplicity. However, this may change in the future
in case we need more such utilities.