refactor ZipperInputStream, create ZipperOutputStream
The "legacy" class `ZipperInputStream` is maintained in
order to keep up the same API as the previous release.

The new class `ZipperOutputStream` uses the same API
as the standard Java class java.util.zip.ZipOutputStream
and can be used as a drop-in replacement.

However, more functions to `ZipperOutputStream` will unleash
the real power to the lazy kind of zipping. Only when passing
files or raw `InputStream` instances to the zipper class,
it can enroll its real power and create ZIP without the use
of any temporary file.