enhance ZipEntry to delete file in fully read, if indicated
deleting temporary files are hard to perform with correct
timing. The best time to do so is, when all the bytes
are fully read and the file is not needed anymore.

Reading the file is subject to the `ZipEntry` class. So
this class is enhanced to provide a flag when setting a file
that will indicate the file is to be deleted when fully read.

This new feature can be used to delete temporary file after
the content has been read.