refactor: provide transition class for new sub-package layout
The class `ZipperInputStream` is located in the sub-package
"zip" now. This would break binary compatibility for all users
of it. So  this transition class indicates to all users of the library,
to use the new location, as it adds a "deprecated" indicator.

It maintains compatibility with old release, yet indicates to use
the new layout and import from the new location.