fix(packets): AbstractZipPacketBuilder: add checks to UInt conversion + tests
converting signed integer to UINT is a bad idea, because the value will
definitely loose the sign and will be a totally different value when decoded.

Although it "should" not happen that signed values are passed to the
conversion functions, you'll never know. Hence bounds checks are added
to ensure a proper exception is thrown for invalid values.

The range for unsigned integer with 16 bits (UNI16) is from
(including) 0 to (excluding) 65536.

The range for unsigned integer with 32 bits (UINT32) is from
(including) 0 to (excluding) 4294967295.

In case the provided values are out of these bounds, then an exception
of type IllegalArgumentException is thrown.