ZipOutputStream add putMultipleNextEntries() to add an enumeration unchanged
This commit restores original behaviour about handling enumerations.

It is useful to add an enumeration directly as a list of entries to the ZIP
instead of fetching all those entries beforehand.

Previously an enumeration had to be converted to a list of entries and
each entry had to be added separately to the ZIP. This behaviour is not
"lazy", though.

The new behaviour really does a lazy read of the enumeration, just when
the bytes of these entries are needed for the ZIP.